expose roo/swap/dumpdevs and other dump vars; make dumpmag consistant int; pedro ok

+++ sys/arch/alpha/alpha/machdep.c
@@ -1,4 +1,3 @@
-
 /*-
  * Copyright (c) 1998, 1999 The NetBSD Foundation, Inc.
  * All rights reserved.
@@ -1104,7 +1103,7 @@ haltsys:
 /*
  * These variables are needed by /sbin/savecore
  */
-u_long	dumpmag = 0x8fca0101;	/* magic number */
+int	dumpmag = 0x8fca0101;	/* magic number */
 int 	dumpsize = 0;		/* pages */
 long	dumplo = 0; 		/* blocks */
